Fix bug in how hidden_twins_uses is counted.
I am using Python 3.6, in which dictionary keys are stored and iterated over in the order
in which they are inserted. Since I insert digits into digits_index according to the order
in which they appear in a box (which at first is in sorted order ascending), the way I
counted hidden_twins_uses did not crop up as a bug during my development and testing; that is,
the invariant (keeping digits in a box in sorted order) just happened to be preserved by the
behavior of Python 3.6. However, my code was implicitly relying on this behavior, so this bug
could manifest in other Python versions.

See: https://docs.python.org/3.6/whatsnew/3.6.html#new-dict-implementation